+++ src/benchmark/fixtures/mod.rs
@@ -1,0 +1,4 @@
+pub const BENCHMARK_COMMON: &str = include_str!("benchmark_common.py");
+pub const BENCHMARK_ACTIONS: &str = include_str!("benchmark_actions.py");
+pub const BENCHMARK_INSTANCES: &str = include_str!("benchmark_instances.py");
+pub const INIT_PY: &str = include_str!("__init__.py");
