+++ src/benchmark/instances.rs
@@ -1,6 +1,6 @@
 use std::{
     env, fs,
-    path::{Path, PathBuf},
+    path::PathBuf,
     process::Command,
     sync::Arc,
     time::{Duration, Instant},
@@ -17,7 +17,10 @@ use tracing::{Instrument, info, warn};
 
 use crate::{
     WorkflowVersionId,
-    benchmark::common::{BenchmarkResult, BenchmarkSummary, spawn_completion_worker},
+    benchmark::{
+        common::{BenchmarkResult, BenchmarkSummary, spawn_completion_worker},
+        fixtures,
+    },
     db::{CompletionRecord, Database},
     instances,
     messages::{MessageError, proto::WorkflowRegistration},
@@ -70,18 +73,23 @@ impl WorkflowBenchmarkHarness {
     ) -> Result<Self> {
         let temp_dir = TempDir::new().context("create temp dir for workflow benchmark")?;
         let repo_root = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-        let fixtures_dir = repo_root.join("src/fixtures");
         let package_dir = temp_dir.path().join("workflow_bench");
         fs::create_dir_all(&package_dir).context("create workflow package dir")?;
-        fs::write(package_dir.join("__init__.py"), b"").context("write package __init__")?;
-        copy_fixture(
-            &fixtures_dir.join("benchmark_common.py"),
-            &package_dir.join("benchmark_common.py"),
-        )?;
-        copy_fixture(
-            &fixtures_dir.join("benchmark_instances.py"),
-            &package_dir.join("benchmark_instances.py"),
-        )?;
+        fs::write(
+            package_dir.join("__init__.py"),
+            fixtures::INIT_PY.trim_start().as_bytes(),
+        )
+        .context("write package __init__")?;
+        fs::write(
+            package_dir.join("benchmark_common.py"),
+            fixtures::BENCHMARK_COMMON.trim_start().as_bytes(),
+        )
+        .context("write benchmark_common fixture")?;
+        fs::write(
+            package_dir.join("benchmark_instances.py"),
+            fixtures::BENCHMARK_INSTANCES.trim_start().as_bytes(),
+        )
+        .context("write benchmark_instances fixture")?;
         let user_module = "workflow_bench.benchmark_instances".to_string();
 
         let python_paths = vec![
@@ -244,11 +252,6 @@ impl WorkflowBenchmarkHarness {
     pub fn actions_per_instance(&self) -> usize {
         self.dag_node_count
     }
-}
-
-fn copy_fixture(src: &Path, dst: &Path) -> Result<()> {
-    fs::copy(src, dst).with_context(|| format!("copy {:?} to {:?}", src, dst))?;
-    Ok(())
 }
 
 fn build_registration_payload(python_paths: &[PathBuf], module: &str) -> Result<Vec<u8>> {
