+++ src/benchmark/mod.rs
@@ -1,3 +1,4 @@
 pub mod actions;
 pub mod common;
+pub mod fixtures;
 pub mod instances;
