+++ src/integration_tests/fixtures/integration_module.py
@@ -1,0 +1,14 @@
+from carabiner_worker import action, workflow
+from carabiner_worker.workflow import Workflow
+
+
+@action
+async def greet(name: str) -> str:
+    return f"hello {name}"
+
+
+@workflow
+class IntegrationWorkflow(Workflow):
+    async def run(self):
+        result = await greet(name="world")
+        return result
